triton/sources/requirements.txt
fastapi>=0.85.0
uvicorn>=0.19.0
pillow>=9.3.0
numpy>=1.23.5
onnxruntime>=1.13.1
python-multipart>=0.0.6
scikit-learn>=1.0.2
